mcp-python-server-client/pyproject.toml
[project]
name = "ai-agents-mcp"
version = "0.1.0"
description = "Demo examples to use for AI Agents with MCP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.25",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-openai>=0.3.21",
    "langgraph>=0.4.8",
    "mcp[cli]>=1.9.3",
    "python-dotenv>=1.1.0",
]
